# apps/web/types/api/common.api.ts
export interface ApiResponse<T> {
  timestamp: string;
  status: number;
  message: string;
  data: T;
}

export interface PaginationMeta {
  page: number;
  limit: number;
  total: number;
  totalPages: number;
  hasNext: boolean;
  hasPrev: boolean;
}

export interface PageResponse<T> {
  data: T[];
  pagination: PaginationMeta;
}