# services/media-service/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "BidNow API",
    "description": "Documentation for BidNow Auction System APIs. \n\nAccess all microservices through the API Gateway. \nAuthentication is handled via JWT Bearer tokens.",
    "contact": {
      "name": "BidNow Support",
      "email": "support@bidnow.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0"
    },
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/v1/media/upload": {
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Upload a file (server-side)",
        "description": "The file passes through the media-service for validation, resizing, and then is uploaded to S3. Allowed types: image/jpeg, image/png, image/webp. Max size: 10 MB.",
        "operationId": "uploadFile",
        "parameters": [
          {
            "name": "entityId",
            "in": "query",
            "description": "Optional: ID of the entity this file belongs to (e.g. auctionId)",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "entityType",
            "in": "query",
            "description": "Optional: Type of the entity (e.g. USER_AVATAR, AUCTION_ITEM)",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "USER_AVATAR",
                "AUCTION_ITEM"
              ]
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "description": "Image file to upload",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Invalid file type or size",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseMediaUploadResponse"
                }
              }
            }
          },
          "201": {
            "description": "File uploaded successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/MediaUploadResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseMediaUploadResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/email-logs/{id}/retry": {
      "post": {
        "tags": [
          "Admin Email Logs"
        ],
        "summary": "Retry email delivery",
        "description": "Manually triggers a retry for a failed email log entry.",
        "operationId": "retryEmail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "UUID of the email log entry",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "550e8400-e29b-41d4-a716-446655440000"
          }
        ],
        "responses": {
          "200": {
            "description": "Retry initiated successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          },
          "404": {
            "description": "Log entry not found",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponseString"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/media/presigned-url": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Generate a presigned upload URL (client-side direct upload)",
        "description": "Returns a short-lived S3 presigned PUT URL. The client uploads the file directly to S3 using this URL — the file never passes through the media-service server. URL is valid for 15 minutes.",
        "operationId": "getPresignedUploadUrl",
        "parameters": [
          {
            "name": "fileName",
            "in": "query",
            "description": "Original file name (used to derive the S3 extension)",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "in": "query",
            "description": "MIME type of the file, e.g. image/png",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Presigned URL generated",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "400": {
            "description": "Unsupported content type",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePresignedUrlResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePresignedUrlResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/email-logs": {
      "get": {
        "tags": [
          "Admin Email Logs"
        ],
        "summary": "Get email logs",
        "description": "Returns a paginated list of email delivery logs.",
        "operationId": "getEmailLogs",
        "parameters": [
          {
            "name": "criteria",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/EmailLogCriteria"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "createdAt"
            }
          },
          {
            "name": "sortDir",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Logs fetched successfully",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageResponseEmailLogResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/BaseResponsePageResponseEmailLogResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "MediaUploadResponse": {
        "required": [
          "contentType",
          "createdAt",
          "fileSize",
          "id",
          "originalName",
          "s3Key"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the media record",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "originalName": {
            "type": "string",
            "description": "Original name of the uploaded file",
            "example": "profile.jpg"
          },
          "s3Key": {
            "type": "string",
            "description": "S3 key of the stored file",
            "example": "media/550e8400-e29b-41d4-a716-446655440000/profile.jpg"
          },
          "contentType": {
            "type": "string",
            "description": "MIME type of the file",
            "example": "image/jpeg"
          },
          "fileSize": {
            "type": "integer",
            "description": "Size of the file in bytes",
            "format": "int64",
            "example": 102400
          },
          "width": {
            "type": "integer",
            "description": "Width of the image (if applicable)",
            "format": "int32",
            "example": 800
          },
          "height": {
            "type": "integer",
            "description": "Height of the image (if applicable)",
            "format": "int32",
            "example": 600
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the media was uploaded",
            "format": "date-time"
          }
        },
        "description": "Response containing uploaded media metadata"
      },
      "BaseResponseMediaUploadResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response",
            "format": "date-time",
            "example": "2024-03-20T10:00:00Z"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/MediaUploadResponse"
          }
        },
        "description": "Generic API response wrapper"
      },
...
      "BaseResponseString": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response",
            "format": "date-time",
            "example": "2024-03-20T10:00:00Z"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Success"
          },
          "data": {
            "type": "string",
            "description": "Response data"
          }
        },
        "description": "Generic API response wrapper"
      },
...
      "PresignedUrlResponse": {
        "required": [
          "expiresAt",
          "s3Key",
          "uploadUrl"
        ],
        "type": "object",
        "properties": {
          "uploadUrl": {
            "type": "string",
            "description": "Presigned URL to upload the file to S3",
            "example": "https://bidnow-media.s3.amazonaws.com/..."
          },
          "s3Key": {
            "type": "string",
            "description": "S3 key for the uploaded file",
            "example": "uploads/uuid/image.jpg"
          },
          "expiresAt": {
            "type": "string",
            "description": "Timestamp when the presigned URL expires",
            "format": "date-time",
            "example": "2023-10-27T10:15:00Z"
          }
        },
        "description": "Presigned URL for media upload"
      },
      "BaseResponsePresignedUrlResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response",
            "format": "date-time",
            "example": "2024-03-20T10:00:00Z"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PresignedUrlResponse"
          }
        },
        "description": "Generic API response wrapper"
      },
...
      "PageResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "List of data items",
            "items": {
              "type": "object",
              "description": "List of data items"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "description": "Paginated response wrapper"
      },
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Current page number (zero-based)",
            "format": "int32",
            "example": 0
          },
          "limit": {
            "type": "integer",
            "description": "Number of items per page",
            "format": "int32",
            "example": 10
          },
          "total": {
            "type": "integer",
            "description": "Total number of items across all pages",
            "format": "int64",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "description": "Total number of pages",
            "format": "int32",
            "example": 10
          },
          "hasNext": {
            "type": "boolean",
            "description": "Whether there is a next page",
            "example": true
          },
          "hasPrev": {
            "type": "boolean",
            "description": "Whether there is a previous page",
            "example": false
          }
        },
        "description": "Metadata for paginated results"
      },
...
      "EmailLogCriteria": {
        "type": "object",
        "properties": {
          "recipientEmail": {
            "type": "string",
            "description": "Filter by recipient email",
            "example": "user@example.com"
          },
          "templateNames": {
            "type": "array",
            "description": "Filter by template names",
            "example": [
              "WELCOME_EMAIL"
            ],
            "items": {
              "type": "string",
              "description": "Filter by template names",
              "example": "[\"WELCOME_EMAIL\"]"
            }
          },
          "statuses": {
            "type": "array",
            "description": "Filter by statuses",
            "example": [
              "SENT",
              "FAILED"
            ],
            "items": {
              "type": "string",
              "description": "Filter by statuses",
              "example": "[\"SENT\",\"FAILED\"]"
            }
          },
          "search": {
            "type": "string",
            "description": "Search term for subject",
            "example": "Welcome"
          }
        },
        "description": "Criteria for searching and filtering email logs"
      },
      "BaseResponsePageResponseEmailLogResponse": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "description": "Timestamp of the response",
            "format": "date-time",
            "example": "2024-03-20T10:00:00Z"
          },
          "status": {
            "type": "integer",
            "description": "HTTP status code",
            "format": "int32",
            "example": 200
          },
          "message": {
            "type": "string",
            "description": "Response message",
            "example": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseEmailLogResponse"
          }
        },
        "description": "Generic API response wrapper"
      },
      "EmailLogResponse": {
        "required": [
          "createdAt",
          "id",
          "recipientEmail",
          "retryCount",
          "status",
          "subject",
          "templateName",
          "updatedAt"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique identifier of the email log",
            "format": "uuid",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "notificationId": {
            "type": "string",
            "description": "Associated notification identifier",
            "format": "uuid",
            "example": "660e8400-e29b-41d4-a716-446655440001"
          },
          "recipientEmail": {
            "type": "string",
            "description": "Email address of the recipient",
            "example": "user@example.com"
          },
          "subject": {
            "type": "string",
            "description": "Subject of the email sent",
            "example": "Welcome to BidNow!"
          },
          "templateName": {
            "type": "string",
            "description": "Name of the template used",
            "example": "WELCOME_EMAIL"
          },
          "status": {
            "type": "string",
            "description": "Current status of the email delivery",
            "example": "SENT"
          },
          "failureReason": {
            "type": "string",
            "description": "Reason for failure if status is FAILED",
            "example": "Invalid recipient address"
          },
          "retryCount": {
            "type": "integer",
            "description": "Number of retries attempted",
            "format": "int32",
            "example": 0
          },
          "sentAt": {
            "type": "string",
            "description": "Timestamp when the email was actually sent",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "description": "Timestamp when the log was created",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "description": "Timestamp when the log was last updated",
            "format": "date-time"
          }
        },
        "description": "Email log details response"
      },
      "PageResponseEmailLogResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "List of data items",
            "items": {
              "$ref": "#/components/schemas/EmailLogResponse"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "description": "Paginated response wrapper"
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}